+++ shhh/scheduler/tasks.py
@@ -7,7 +7,6 @@ from shhh.models import Entries
 def delete_expired_links():
     """Delete expired links from the database."""
     app = scheduler.app
-
     with app.app_context():
         db.session.query(Entries).filter(
             Entries.date_expires <= datetime.now()).delete()
